requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
abseil-cpp=20211102.0=h27087fc_1
aiohttp=3.8.1=py38h0a891b7_1
aiosignal=1.3.1=pyhd8ed1ab_0
arrow-cpp=8.0.0=py38h60b952e_0
async-timeout=4.0.2=pyhd8ed1ab_0
attrs=22.2.0=pyh71513ae_0
av=10.0.0=pypi_0
aws-c-common=0.4.57=he6710b0_1
aws-c-event-stream=0.1.6=h2531618_5
aws-checksums=0.1.9=he6710b0_0
aws-sdk-cpp=1.8.185=hce553d0_0
blas=1.0=mkl
boost-cpp=1.78.0=he72f1d9_0
bottleneck=1.3.5=py38h7deecbd_0
brotli=1.0.9=h166bdaf_7
brotli-bin=1.0.9=h166bdaf_7
brotlipy=0.7.0=py38h0a891b7_1004
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f98852_0
ca-certificates=2022.12.7=ha878542_0
certifi=2022.12.7=pypi_0
cffi=1.15.0=py38h3931269_0
charset-normalizer=2.1.1=pyhd8ed1ab_0
click=8.1.3=unix_pyhd8ed1ab_2
colorama=0.4.6=pyhd8ed1ab_0
cryptography=3.4.8=py38h3e25421_1
cuda=11.7.1=0
cuda-cccl=11.7.91=0
cuda-command-line-tools=11.7.1=0
cuda-compiler=11.7.1=0
cuda-cudart=11.7.99=0
cuda-cudart-dev=11.7.99=0
cuda-cuobjdump=11.7.91=0
cuda-cupti=11.7.101=0
cuda-cuxxfilt=11.7.91=0
cuda-demo-suite=12.0.76=0
cuda-documentation=12.0.76=0
cuda-driver-dev=11.7.99=0
cuda-gdb=12.0.90=0
cuda-libraries=11.7.1=0
cuda-libraries-dev=11.7.1=0
cuda-memcheck=11.8.86=0
cuda-nsight=12.0.78=0
cuda-nsight-compute=12.0.0=0
cuda-nvcc=11.7.99=0
cuda-nvdisasm=12.0.76=0
cuda-nvml-dev=11.7.91=0
cuda-nvprof=12.0.90=0
cuda-nvprune=11.7.91=0
cuda-nvrtc=11.7.99=0
cuda-nvrtc-dev=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-nvvp=12.0.90=0
cuda-runtime=11.7.1=0
cuda-sanitizer-api=12.0.90=0
cuda-toolkit=11.7.1=0
cuda-tools=11.7.1=0
cuda-visual-tools=11.7.1=0
dataclasses=0.8=pyhc8e2a94_3
datasets=2.9.0=pyhd8ed1ab_0
dill=0.3.6=pyhd8ed1ab_1
ffmpeg=4.2.3=h167e202_0
filelock=3.9.0=pyhd8ed1ab_0
flit-core=3.6.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.3.3=py38h5eee18b_0
fsspec=2023.1.0=pyhd8ed1ab_0
fvcore=0.1.5.post20221221=pypi_0
gds-tools=1.5.0.59=0
gflags=2.2.2=he1b5a44_1004
glog=0.6.0=h6f12383_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
grpc-cpp=1.46.1=h33aed49_0
huggingface-hub=0.14.1=pypi_0
icu=70.1=h27087fc_0
idna=3.4=pyhd8ed1ab_0
importlib-metadata=6.0.0=pyha770c72_0
importlib_metadata=6.0.0=hd8ed1ab_0
intel-openmp=2021.4.0=h06a4308_3561
iopath=0.1.10=pypi_0
joblib=1.2.0=pyhd8ed1ab_0
keras-preprocessing=1.1.2=pypi_0
keyutils=1.6.1=h166bdaf_0
krb5=1.19.3=h3790be6_0
lame=3.100=h7b6447c_0
ld_impl_linux-64=2.38=h1181459_1
libblas=3.9.0=12_linux64_mkl
libbrotlicommon=1.0.9=h166bdaf_7
libbrotlidec=1.0.9=h166bdaf_7
libbrotlienc=1.0.9=h166bdaf_7
libcblas=3.9.0=12_linux64_mkl
libcublas=11.10.3.66=0
libcublas-dev=11.10.3.66=0
libcufft=10.7.2.124=h4fbf590_0
libcufft-dev=10.7.2.124=h98a8f43_0
libcufile=1.5.0.59=0
libcufile-dev=1.5.0.59=0
libcurand=10.3.1.50=0
libcurand-dev=10.3.1.50=0
libcurl=7.87.0=h91b91d3_0
libcusolver=11.4.0.1=0
libcusolver-dev=11.4.0.1=0
libcusparse=11.7.4.91=0
libcusparse-dev=11.7.4.91=0
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libevent=2.1.10=h9b69904_4
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=12.2.0=h69a702a_19
libgfortran5=12.2.0=h337968e_19
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
liblapack=3.9.0=12_linux64_mkl
libnghttp2=1.46.0=hce63b2e_0
libnpp=11.7.4.75=0
libnpp-dev=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libnvjpeg-dev=11.8.0.2=0
libpng=1.6.37=hbc83047_0
libprotobuf=3.20.1=h4ff587b_0
libssh2=1.10.0=ha56f1ee_2
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libthrift=0.15.0=he6d91bd_0
libunistring=0.9.10=h27cfd23_0
lz4-c=1.9.3=h9c3ff4c_1
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
multidict=6.0.2=py38h5eee18b_0
multiprocess=0.70.12.2=py38h0a891b7_2
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.0=pypi_0
nsight-compute=2022.4.0.15=0
numexpr=2.8.4=py38he184ba9_0
numpy=1.23.5=py38h14f4228_0
numpy-base=1.23.5=py38h31eccc5_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.1.23=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1s=h7f8727e_0
orc=1.7.4=h07ed6aa_0
packaging=23.0=pyhd8ed1ab_0
pandas=1.5.2=py38h417a72b_0
parameterized=0.8.1=pypi_0
pip=22.3.1=py38h06a4308_0
portalocker=2.7.0=pypi_0
pyarrow=8.0.0=py38h992f0b0_0
pycparser=2.21=pyhd8ed1ab_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pysocks=1.7.1=pyha2e5f31_6
python=3.8.16=h7a1cb2a_2
python-dateutil=2.8.2=pyhd8ed1ab_0
python-xxhash=3.0.0=py38h0a891b7_1
python_abi=3.8=2_cp38
pytorch=1.13.1=py3.8_cuda11.7_cudnn8.5.0_0
pytorch-cuda=11.7=h67b0de4_1
pytorch-mutex=1.0=cuda
pytorchvideo=0.1.5=pypi_0
pytz=2022.7.1=pyhd8ed1ab_0
pyyaml=6.0=py38h0a891b7_4
re2=2022.04.01=h27087fc_0
readline=8.2=h5eee18b_0
regex=2022.7.9=py38h5eee18b_0
requests=2.28.2=pyhd8ed1ab_0
responses=0.18.0=pyhd8ed1ab_0
sacremoses=0.0.53=pyhd8ed1ab_0
scikit-learn=1.0.2=py38h1561384_0
scipy=1.8.1=py38h1ee437e_0
setuptools=65.6.3=py38h06a4308_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.9=hbd366e4_1
sqlite=3.40.1=h5082296_0
tabulate=0.9.0=pypi_0
tensorrt=8.6.1=pypi_0
tensorrt-bindings=8.6.1=pypi_0
tensorrt-libs=8.6.1=pypi_0
termcolor=2.2.0=pypi_0
threadpoolctl=3.1.0=pyh8a188c0_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.11.4=py38h3dcd8bd_1
torchaudio=0.13.1=py38_cu117
torchmetrics=0.11.0=pyhd8ed1ab_0
torchvision=0.15.0a0+78ffda7=pypi_0
tqdm=4.64.1=pyhd8ed1ab_0
transformers=4.30.0.dev0=pypi_0
typing-extensions=4.4.0=py38h06a4308_0
typing_extensions=4.4.0=py38h06a4308_0
urllib3=1.26.14=pyhd8ed1ab_0
utf8proc=2.6.1=h27cfd23_0
wheel=0.37.1=pyhd3eb1b0_0
x264=1!152.20180806=h7b6447c_0
xxhash=0.8.0=h7f98852_3
xz=5.2.10=h5eee18b_1
yacs=0.1.8=pypi_0
yaml=0.2.5=h7f98852_2
yarl=1.7.2=py38h0a891b7_2
zipp=3.12.0=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
